+++ actix-limitation/CHANGES.md
@@ -3,9 +3,11 @@
 ## Unreleased - 2022-xx-xx
 - Update Actix Web dependency to v4 ecosystem. [#229]
 - Update Tokio dependencies to v1 ecosystem. [#229]
-- 
+- Rename `Limiter::{build => builder}()`. [#???]
+- Rename `Builder::{finish => build}()`. [#???]
 
 [#229]: https://github.com/actix/actix-extras/pull/229
+[#???]: https://github.com/actix/actix-extras/pull/???
 
 
 ## 0.1.4 - 2022-03-18